+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "musae",
-  "version": "0.1.23-beta.4",
+  "version": "0.1.24",
   "description": "musae-ui",
   "author": "tutu@fantufantu.com",
   "license": "MIT",
